# swagger.yaml
openapi: 3.0.3
info:
  title: VideoJuegos API
  description: API en Go para buscar juegos en RAWG y administrar una biblioteca personal en MongoDB.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
...
  /api/library:
    get:
      summary: Lista juegos guardados en la biblioteca
      parameters:
        - name: status
          in: query
          description: Filtrar por estado (pendiente, jugando, completado, abandonado)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de juegos guardados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Agrega un juego a la biblioteca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rawg_id
                - title
              properties:
                rawg_id:
                  type: integer
                title:
                  type: string
                genre:
                  type: string
                platform:
                  type: string
                cover_url:
                  type: string
      responses:
        '201':
          description: Juego creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Conflicto - rawg_id ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    Game:
      type: object
      properties:
        id:
          type: string
        rawg_id:
          type: integer
        title:
          type: string
        genre:
          type: string
        platform:
          type: string
        cover_url:
          type: string
        personal_note:
          type: string
        personal_score:
          type: integer
          minimum: 1
          maximum: 10
        status:
          type: string
          enum:
            - pendiente
            - jugando
            - completado
            - abandonado
        added_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        error:
          type: string
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'